# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Majstorce API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://api.majstorce.mk/api/latest",
            "description": "Production Enviroment"
        },
        {
            "url": "http://api.majstorce.mk.local/api/latest",
            "description": "Local Enviroment - with host"
        },
        {
            "url": "http://localhost/api/latest",
            "description": "Local Enviroment -  docker"
        }
    ],
    "paths": {
...
        "/oauth/refresh-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Exchange refresh token for a new access token",
                "operationId": "RefreshTokenExchange::refreshTokenExchange_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestRefreshTokenExchange"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration was saccesfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseAuthorizationTokens"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Precondition failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/oauth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Cliant Registraion",
                "operationId": "Register::register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestClientRegister"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration was saccesfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseAuthorizationTokens"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Precondition failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/oauth/reset-password/request": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Reset Password",
                "operationId": "ResetPassword::request_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestResetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email for reset password is sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseSuccess"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Precondition failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/oauth/reset-password/new-rassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Reset Password With Token",
                "operationId": "ResetPassword::newPasswordWithToken_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestResetPasswordWithToken"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucessfully update password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseSuccess"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Precondition failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorizati3on",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}